+++ packages/visual-stack/src/components/tests/Form.test.js
@@ -4,7 +4,14 @@ import Adapter from 'enzyme-adapter-react-16';
 
 Enzyme.configure({ adapter: new Adapter() });
 
-import { ChoiceInput, Field, Input, TextField, SelectField } from '../Form';
+import {
+  ChoiceInput,
+  Field,
+  Input,
+  TextArea,
+  TextField,
+  SelectField,
+} from '../Form';
 
 describe('Form', () => {
   describe('Field', () => {
@@ -47,6 +54,24 @@ describe('Form', () => {
       );
 
       wrapper.find('input').simulate('blur', { target: { value: '  test  ' } });
+
+      expect(mockOnChange.mock.calls[0][0].target.value).toEqual('test');
+      expect(mockOnBlur.mock.calls[0][0].target.value).toEqual('test');
+    });
+  });
+
+  describe('TextArea', () => {
+    it('should trim on blur', () => {
+      const mockOnChange = jest.fn();
+      const mockOnBlur = jest.fn();
+
+      const wrapper = mount(
+        <TextArea trimmed onBlur={mockOnBlur} onChange={mockOnChange} />
+      );
+
+      wrapper
+        .find('textarea')
+        .simulate('blur', { target: { value: '  test  ' } });
 
       expect(mockOnChange.mock.calls[0][0].target.value).toEqual('test');
       expect(mockOnBlur.mock.calls[0][0].target.value).toEqual('test');