+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Upcoming
+## Updates
+- TextArea: Add trimmed prop to trim value on blur
+
 # 6.7.5
 ## Updates
 - Input: Add trimmed prop to trim value on blur